doc: extract “Neuropil from the Bottom Up” chapter (wip)
The idea here is that explaining and understanding Neuropil will be much easier
“bottom up” than “top down”. Another rationale is that early adopters of
Neuropil will most likely be “power users” and benefit from in-depth resources.

Further work will tell if this approach is feasible, since we can’t possibly
cover all things in depth. In order for this to work we need to find a good
intersection of tapestry and Neuropil’s crypto architecture, that makes sense
when presented semi-superficially and won’t blow your mind, but still covers
everything you need to know in order for Neuropil’s higher level abstractions
to come naturally.