Use plain np_id instead of np_id * in API
np_id is an array, np_id * is a pointer to an array which happens to be the
same address as the address of the first element of the array... C is tricky.

The motivation for this is to simplify the interface. While it is trivial
to get the pointer to an array in C, it might be less so in foreign languages
like when using the LuaJIT FFI where ffi.new("np_id") will yield an np_id while
having no concept of a pointer to the resulting object.

Internally this complicates things a bit: C does not allow casting to an array
type. So to convert from the legacy np_dhkey_s we need to do the following
dance:

  struct np_dhkey_s *k;
  *(np_id *)&k // yields a np_id

I.e., we cast a pointer to struct np_dhkey_s to a pointer to np_id which we
then dereference to get an np_id. All to avoid the forbidden "cast to array."
Open question: is this even defined behavior? It’s definitely not pretty.

Along the way:

 - I removed the mistaken application context parameter from np_get_id (pure
   function from data to fingerprint)

 - np_id2str/str2id became np_str_id/id_str in canonical form (destination
   buffer as their first argument). As to the naming scheme: I find it
   confusing to occasionally introduce digit/word synonyms. Was it id_to_str?
   id2str? id_str!