Constify read-only string parameters
This is relevant to the FFI because it signals host languages like Lua that
they can safely pass their native strings into Neuropil FFI calls. It’s also
good practice to do so in any case. On that note: we should investigate all
warnings emitted by the compiler that due to const. These messages look like
this:

 warning: passing 'const uint8_t *' (aka 'const unsigned char *') to parameter
 of type 'unsigned char *' discards qualifiers

We should follow them to ensure that const qualifiers are not discarded, which
will possibly lead us to bugs where we mistakenly modify const arguments.